requirements.txt
Appium-Python-Client==0.38
Faker==1.0.2
python-dateutil==2.8.0
PyYAML
selenium
pytest
pip
six==1.12.0
text-unidecode==1.2
urllib3==1.24.1
